+++ src/components/button/button.component.tsx
@@ -15,6 +15,7 @@ import { TooltipPositions } from "../tooltip/tooltip.config";
 import ButtonBarContext from "../button-bar/__internal__/button-bar.context";
 import SplitButtonContext from "../split-button/__internal__/split-button.context";
 import BatchSelectionContext from "../batch-selection/__internal__/batch-selection.context";
+import ButtonContext from "./__internal__/button.context";
 
 export type ButtonTypes =
   | "primary"
@@ -120,6 +121,7 @@ interface RenderChildrenProps
     | "buttonType"
     | "iconTooltipMessage"
     | "iconTooltipPosition"
+    | "destructive"
   > {
   buttonType: ButtonTypes;
   tooltipTarget?: HTMLElement;
@@ -136,6 +138,7 @@ function renderChildren({
   iconTooltipMessage,
   iconTooltipPosition,
   tooltipTarget,
+  destructive,
 }: RenderChildrenProps) {
   const iconColor = () => {
     if (buttonType === "primary") {
@@ -169,7 +172,15 @@ function renderChildren({
       {isValidChildren && (
         <span>
           <StyledButtonMainText data-element="main-text">
-            {children}
+            <ButtonContext.Provider
+              value={{
+                isInsideTypicalButton: !destructive,
+                isInsideDestructiveButton: destructive,
+                isInsidePrimaryButton: buttonType === "primary",
+              }}
+            >
+              {children}
+            </ButtonContext.Provider>
           </StyledButtonMainText>
           {size === "large" && (
             <StyledButtonSubtext data-element="subtext" data-role="subtext">
@@ -353,6 +364,7 @@ const Button = React.forwardRef<
           buttonType,
           iconTooltipMessage,
           iconTooltipPosition,
+          destructive,
           tooltipTarget: internalRef as HTMLElement | undefined,
         })}
       </StyledButton>