+++ src/components/vertical-menu/responsive-vertical-menu/components.test-pw.tsx
@@ -195,6 +195,11 @@ export const WithDifferentDepthsAsLastItem = () => {
             </ResponsiveVerticalMenuItem>
           </ResponsiveVerticalMenuItem>
         </ResponsiveVerticalMenuItem>
+        <ResponsiveVerticalMenuItem
+          id="primary-last-menu-item"
+          label="Last Primary Menu Item"
+          onClick={() => {}}
+        />
       </ResponsiveVerticalMenu>
     </Box>
   );
@@ -239,3 +244,40 @@ export const WithLongList = () => {
     </ResponsiveVerticalMenu>
   );
 };
+
+export const WithSiblingControl = () => (
+  <Box height="100vh">
+    <ResponsiveVerticalMenu height="100%">
+      <ResponsiveVerticalMenuItem
+        icon="home"
+        id="primary-menu"
+        label="Primary Menu With Children"
+      >
+        <ResponsiveVerticalMenuItem
+          id="secondary-menu"
+          label="Secondary Menu With Children"
+        >
+          <ResponsiveVerticalMenuItem
+            id="tertiary-menu"
+            label="Tertiary Menu"
+          />
+        </ResponsiveVerticalMenuItem>
+        <ResponsiveVerticalMenuItem
+          id="secondary-menu-no-children"
+          label="Secondary Menu Item"
+        />
+      </ResponsiveVerticalMenuItem>
+      <ResponsiveVerticalMenuDivider />
+      <ResponsiveVerticalMenuItem
+        icon="home"
+        id="primary-menu-no-children"
+        label="Primary Menu Item"
+        onClick={() => {}}
+      />
+    </ResponsiveVerticalMenu>
+    <label>
+      Check me!
+      <input type="checkbox" />
+    </label>
+  </Box>
+);